backend/requirements.txt
Flask
Flask-Cors
aiohttp
python-dotenv
requests
beautifulsoup4
lxml
urllib3